File IO class and subclasses
