+++ utils/csv_io.py
@@ -1,29 +1,9 @@
-import s3fs
 import pandas as pd
-import os
+from utils.file_io import FileIO
 
 
-class IOcsv:
-    def get_credentials(self):
-        '''
-        Get the credentials to connect to the S3 Datalab file system
-        '''
-        # environment variables
-        CLIENT_KWARGS = {
-            'endpoint_url': 'https://' + os.environ['AWS_S3_ENDPOINT']
-        }
-        KEY = os.environ['AWS_ACCESS_KEY_ID']
-        SECRET = os.environ['AWS_SECRET_ACCESS_KEY']
-        TOKEN = os.environ['AWS_SESSION_TOKEN']
-        file_system = s3fs.S3FileSystem(
-            client_kwargs=CLIENT_KWARGS,
-            key=KEY,
-            secret=SECRET,
-            token=TOKEN
-        )
-        return file_system
-
-    def import_csv(
+class IOcsv(FileIO):
+    def import_file(
         self, bucket: str, file_key_s3: str,
         encoding='utf8', sep=';'
     ):
@@ -40,7 +20,7 @@ class IOcsv:
             df = pd.read_csv(file_input, sep=sep, encoding=encoding)
         return df
 
-    def export_csv(self, df, bucket: str, file_key_s3: str):
+    def export_file(self, df, bucket: str, file_key_s3: str):
         '''
         export_csv allow to transform a pandas dataframe into a csv file
         and to export it with any vscode service on the datalab using
@@ -49,5 +29,5 @@ class IOcsv:
         file_system = self.get_credentials()
         file_path_s3 = bucket + "/" + file_key_s3
 
-        with file_system.open(file_path_s3, 'w') as file_output:
+        with file_system.open(file_path_s3, mode='w') as file_output:
             df.to_csv(file_output, index=False, sep=';')
